check_builds/pyproject.toml
[project]
name = "check-builds"
version = "0.1.0"
description = "Check that builds produced desired artifacts"
authors = [
    { name = "Couchbase Build Team", email = "build-team@couchbase.com" }
]
dependencies = ["Jinja2", "PyYAML", "cbbuild-database>=0.8.1"]
readme = "README.md"
requires-python = ">= 3.10, <3.11"

[project.scripts]
check-builds = "check_builds.check_builds_prog:main"
test-yaml = "check_builds.util:main"

[tool.uv]
package = true